Silly bug where internal value for IntValue was stored as a double!

+++ libs/imp/include/imp/runtime/intvalue.hpp
@@ -36,7 +36,7 @@ namespace Imp
         std::shared_ptr<RuntimeValue> EvalStr(ImpSyntax *where) override;
 
     private:
-        double _Value;
+        long _Value;
     };
 }
